pyproject.toml
[project]
name = "langchain-start"
version = "0.1.0"
description = ""
authors = [
    {name = "Germán Küber",email = "german.kuber@outlook.com"}
]
readme = "README.md"
requires-python = "3.12.9"
dependencies = [
    "langchain[openai] (>=0.3.23,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-core (>=0.3.52,<0.4.0)",
    "langgraph (>=0.3.30,<0.4.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "langchain-text-splitters (>=0.3.8,<0.4.0)",
    "ipython (>=9.1.0,<10.0.0)",
    "langsmith (>=0.3.31,<0.4.0)",
    "langchain-anthropic (>=0.3.11,<0.4.0)"
]

[tool.poetry]
package-mode = false

 
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
